test-playlist-manager: do not free the manager
The documentation is clear about it:

    the playlist manager of a process. A process has only one object
    like this, and it must not be released by the caller.

Yet the tests seem to want to free the manager, perhaps to test some
intricate part, but it's clear that this should not be done.

Either the tests needs to be split up into two tests, or we just do not
unref the manager.